e.printStackTrace()

+++ src/ru/vsu/chuprikov/task8/ConsoleApp.java
@@ -8,13 +8,22 @@ import java.io.*;
 public class ConsoleApp {
     public static void main(String[] args) {
         InputArgs inputArgs = ConsoleUtils.parseCmdArgs(args);
-        int[][] matrix = Matrix.readMatrixFromFile(inputArgs.inputFile);
-        printResult(matrix, inputArgs.outputFile);
+        try {
+            int[][] matrix = Matrix.readMatrixFromFile(inputArgs.inputFile);
+            printResult(matrix, inputArgs.outputFile);
+        } catch (Exception e) {
+            e.printStackTrace();
+        }
     }
 
-    public static void printResult(int[][] matrix, String filename) {
+    public static void printResult(int[][] matrix, String filename) throws FileNotFoundException {
         String path = "C:\\Scripts\\Java\\Tasks\\src\\ru\\vsu\\chuprikov\\task8\\";
         File file = new File(path + filename);
+
+        if (!file.exists()) {
+            throw new FileNotFoundException(filename + " not found");
+        }
+
         try (FileWriter fileWriter = new FileWriter(file)) {
             boolean result = Matrix.isMatrixSorted(matrix);
             if (result) {
@@ -22,7 +31,8 @@ public class ConsoleApp {
             } else {
                 fileWriter.write("Элементы матрицы не образуют упорядоченную последовательность.");
             }
+        } catch (Exception e) {
+            e.printStackTrace();
         }
-        catch (Exception e) { }
     }
 }
